Correct typos and add secret keyboard commands
In this commit, various typos in the `Login.vue` file were corrected. A secret key press combination was added to the `Encuesta1.vue` and `Encuesta2.vue` files to automatically populate and submit survey form data. Also, a change in redirect path was made to point to "inicio" instead of "calendario" when an unknown path is accessed. Some elements in navigation bar were re-ordered and title changes were made in `Encuesta1.vue` and `Encuesta2.vue` files.